requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4

# PDF Processing
PyMuPDF==1.24.1
pdfplumber==0.11.0

# NLP
nltk==3.8.1
spacy==3.7.4
scikit-learn==1.4.1.post1

# Keyword Extraction
keybert==0.8.4
rake-nltk==1.0.6

# Summarization & Transformers
transformers==4.40.0
torch==2.2.2
sentencepiece==0.2.0

# Sentence Similarity
sentence-transformers==2.7.0

# Vector Store
faiss-cpu==1.8.0

# Data & Visualization
numpy==1.26.4
pandas==2.2.2
plotly==5.21.0
matplotlib==3.8.4
seaborn==0.13.2
networkx==3.2.1

# Frontend
streamlit==1.33.0

# Utilities
requests==2.31.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0

# Database & Auth
motor==3.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pymongo==4.6.2
